Set up basic timer loop
The time to execute a task is determined by recording the timestamps
before and after, and measuring the difference. Results are summed over
multiple iterations to smooth anomalies.

The commented-out code is from an initial effort. It was noted that
placing the loop between the start and end time recordings meant the
difference would not represent the time to execute only the operation of
interest, but also the time to execute the iteration. However, it is
worth noting that this method returned much faster times than the one
which has been left uncommented, suggesting it was not executing the
operation anew in each iteration (also suggesting the iteration itself
is executed in minimal time compared to reading from the end of the
array).

Also of note is that the first periodic sample consistently returns a
higher delay than the later ones - in general it appears the first batch
of 10000 operations takes around 25% longer to complete than the others,
which progress in an approximately linear manner as expected. This could
be related to page loading - further investigation required!