Extend basic test
Added iterations through arrays containing different data type/varying
size of array entries, and through a very large array.

Although the size of array has no noticeable impact on the times
recorded by the test, creating an array orders of magnitude larger than
1,000,000 had a significant effect on browser performance - the page
would not load when trying to create an array of 100,000,000 entries.

There is a second function which more clearly demonstrates the issue
previously identified, whereby the first timings added to the total are
significantly longer than those run later in the test. There was no
noticeable benefit from adding an eventListener to run the script only
after the page loads; and by disregarding an initial portion of the
timings measured, it is clear the lag is not related to when the script
or function starts running - or at least, not uniquely so - rather it is
evident for the first 20,000-30,000 timings that are added to the total.