Improve tailoring of test parameters
The size of arrays and number of repetitions can now be declared for
each test, so that values can be selected to give best results for the
function of interest.

Charts are now displayed as scatter plots, as the lines produced were
generally not helpful for seeing trends. Google Charts does include
capability for different trendlines, but as the type needs to be
specified in advance, adding these is unlikely to add value for now.

Observed interesting behaviour of array.sort due to the changes
(performance spike around array size 100000) suggesting JS may apply
different sort techniques depending on the data presented.