Refactor to separate out the timing function
Duplicating the existing tests showed that recording the time interval
around a function call containing the operation, does not significantly
affect the outcome compared with combining the operation and the timing
code in one function.

This refactor will allow more general testing of the time efficiency of
functions without having to include timer code in the function passed
when running the test.