Add tests for a shuffle(array) function
The test for values being moved to new positions checks for a range of
differences between the old and new indices of each value, to ensure the
array has not simply been shifted/neighbouring positions swapped etc.
There is a risk that this will occasionally cause the test to fail if
the shuffle is truly random; to mitigate this, it is only expected that
one tenth of the differences will be unique. This test should be kept
under review going forward and ultimately it may be more appropriate to
only test this aspect manually.