Add shuffle function to performance tests
I expected this function to have linear performance based only iterating
over the array once to write a new value to each position, but it is in
fact quadrilateral. On review, it is likely that use of splice is having
an impact by introducing a second 'edit array' operation, although not
yet known whether splice is such a demanding operation that it would be
the significant factor in this regard.