Change performance charts to display mean timings
Comparing charts previously required knowledge of how many repetitions
were specified in the code for each test being run, as the total time to
repeatedly carry out the operation was displayed. This is fine if the
values are being kept identical for the tests which will be compared,
and in general it makes sense to keep the test conditions as similar as
reasonably possible, but switching to displaying a mean allows at least
some comparison of functions that require very different argument values
due to the performance differing by orders of magnitude.