Add a suite of test-driven 'reverse' functions
Some very similar variants have been included for comparison, in
particular those grouped under 'describe "reverse a copy"' in the
tests. Surprisingly, use of array.slice to copy the array was
noticeably more efficient than populating an empty array, even when
using the same technique to overwrite the entries.

Speculation: the process of reversing the order of entries in the array
may complete so quickly that the time taken to dynamically resize an
array which is initially empty becomes significant. Slicing the array
will cut out this resizing by effectively declaring a fixed size for the
copy array when it is assigned.

Screenshots have been added to the images folder which show the
difference in performance when slicing vs starting with an empty array;
of further interest is the difference in performance between Firefox and
Chrome browsers when each is used to run the tests.