Further screenshot added for browser comparison
Chrome outperforms Firefox for all functions that don't involve an empty
array, particularly the 2 functions that only iterate over half the
array (reverse and otherReverseCopy) which run around twice as fast or
better in Chrome. In contrast, the functions which populate an empty
array take around 75% as long to run in Firefox, as seen in the previous
images.

For purposes of this exercise it is not practical to run all tests in a
selection of browsers every time, but this is a nice example of the
challenges inherent in trying to optimise code that will be run on a
number of different platforms.