Add a 'base 3' merge sort function
For an array of size 3, bubble sorting (comparing and swapping adjacent
pairs of values) is quite efficient as the maximum necessary number of
swaps is three - which applies if the array needs reversing completely -
and the best case is zero swaps needed if the array is already sorted.
Using 'divide and conquer' to sort an array of size 3 involves either 2
or 3 actions: split, swap if necessary, and merge, where merging is not
entirely trivial as it involves iteration.

It therefore seemed possible that introducing a bubble sort for arrays
of size 3 could improve the efficiency of the merge sort function. This
would be largely dependent on the relative efficiencies of splitting the
array, swapping two values, and merging two sorted arrays.

Performance tests were run using a variety of base array sizes to catch
any differences when the sorting step happens only on arrays of size 2
(this will be the case when the length of the original array is a power
of 2), only on arrays of size 3 (when the original array length is a
power of two multiplied by 3, i.e. (2^n)*3), or a mix of both. These
showed no significant advantage in adopting either 2-base or 3-base
sorting - both functions produced time efficiency curves that were
effectively identical, allowing for random variation. However, the
additional code complexity of the threeBaseMergeSort function makes it
undesirable for anything other than an academic exercise.

Consideration was also given to making the splitting & merging processes
trinary rather than binary, but this was not implemented as it adds
complexity to the merging: three values rather than two must be compared
to find the minimum on each iteration. It is possible this would be
efficient for edge cases like [27 els] -> 3*[9 els] -> 9*[3 els] versus
[27 els] -> [14 els]+[13 els] ... -> 6*[2 els]+5*[3 els]; but any gain
there, if such exists, is unlikely to be so drastic that it offsets
other weaknesses of the method.